Add files via upload
This code manages a modal's visibility on a webpage. It defines a function, toggleModal, to add or remove the 'open-modal' class from the .modal-overlay element, thereby showing or hiding the modal. The visibility is toggled via event listeners attached to buttons: one to open the modal (modal-btn) and another to close it (close-btn). This approach provides a dynamic way to control modal displays with minimal code, enhancing user interaction without reloading the page.